allow user to send message to chatwoot if operator had contacted them
